requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
mysql-connector-python
requests
